# backend/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/members": {
            "get": {
                "operationId": "MembersController_findAll",
                "summary": "Get info about all members",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MemberDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "post": {
                "operationId": "MembersController_create",
                "summary": "Create new member",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateMemberDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/api/members/{id}": {
            "get": {
                "operationId": "MembersController_findById",
                "summary": "Get full info about member",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "patch": {
                "operationId": "MembersController_update",
                "summary": "Update member",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateMemberDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
...
        "/api/member-transactions": {
            "get": {
                "operationId": "MemberTransactionsController_findAll",
                "summary": "Get all member transactions",
                "parameters": [
                    {
                        "name": "offset",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "count",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderBy",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderDirection",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberTransactionsDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "member-transactions"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "post": {
                "operationId": "MemberTransactionsController_create",
                "summary": "Create new member transaction",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMemberTransactionDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberTransactionDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "member-transactions"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
...
        "/api/space-transactions": {
            "get": {
                "operationId": "SpaceTransactionsController_findAll",
                "summary": "Get all space transactions",
                "parameters": [
                    {
                        "name": "offset",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "count",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderBy",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderDirection",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SpaceTransactionsDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "space-transactions"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "post": {
                "operationId": "SpaceTransactionsController_create",
                "summary": "Create new space transaction",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSpaceTransactionDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SpaceTransactionDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "space-transactions"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/api/space-transactions/actor/{memberId}": {
            "get": {
                "operationId": "SpaceTransactionsController_findAllByActor",
                "summary": "Get all space transactions",
                "parameters": [
                    {
                        "name": "memberId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "count",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderBy",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderDirection",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SpaceTransactionsDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "space-transactions"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "RE: Swynca",
        "description": "B4CKSP4CE member/money management system",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "session"
            }
        },
        "schemas": {
            "TelegramMetadataDTO": {
                "type": "object",
                "properties": {
                    "telegramId": {
                        "type": "string"
                    },
                    "telegramName": {
                        "type": "string"
                    }
                },
                "required": [
                    "telegramId",
                    "telegramName"
                ]
            },
            "GitHubMetadataDTO": {
                "type": "object",
                "properties": {
                    "githubId": {
                        "type": "string"
                    },
                    "githubUsername": {
                        "type": "string"
                    }
                },
                "required": [
                    "githubId",
                    "githubUsername"
                ]
            },
            "MemberDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "frozen"
                        ]
                    },
                    "balance": {
                        "type": "string"
                    },
                    "joinedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "telegramMetadata": {
                        "$ref": "#/components/schemas/TelegramMetadataDTO"
                    },
                    "githubMetadata": {
                        "$ref": "#/components/schemas/GitHubMetadataDTO"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "email",
                    "status",
                    "balance",
                    "joinedAt"
                ]
            },
...
            "ErrorApiResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "statusCode",
                    "message"
                ]
            },
...
            "CreateUpdateMemberDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                },
                "required": [
                    "name",
                    "email"
                ]
            },
...
            "MemberTransactionDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "deposit",
                            "withdrawal"
                        ]
                    },
                    "amount": {
                        "type": "string"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "magic",
                            "donate",
                            "topup"
                        ]
                    },
                    "target": {
                        "type": "string",
                        "enum": [
                            "magic",
                            "membership"
                        ]
                    },
                    "actor": {
                        "$ref": "#/components/schemas/MemberDTO"
                    },
                    "subject": {
                        "$ref": "#/components/schemas/MemberDTO"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "type",
                    "amount",
                    "date",
                    "actor",
                    "subject",
                    "createdAt"
                ]
            },
            "MemberTransactionsDTO": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "number"
                    },
                    "transactions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MemberTransactionDTO"
                        }
                    }
                },
                "required": [
                    "count",
                    "transactions"
                ]
            },
            "CreateMemberTransactionDTO": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "deposit",
                            "withdrawal"
                        ]
                    },
                    "amount": {
                        "type": "string"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "magic",
                            "donate",
                            "topup"
                        ]
                    },
                    "target": {
                        "type": "string",
                        "enum": [
                            "magic",
                            "membership"
                        ]
                    },
                    "subjectId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "type",
                    "amount",
                    "date",
                    "subjectId"
                ]
            },
...
            "SpaceTransactionsDTO": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "number"
                    },
                    "transactions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SpaceTransactionDTO"
                        }
                    }
                },
                "required": [
                    "count",
                    "transactions"
                ]
            },
...
        }
    }
}